loolwsd: Improved loolwsd -> storage communication
The storage object was recreated superfluously everytime a new
view of a document is loaded. This means that with every load, we
forget everything (states etc.) and download the
file from the storage to our chroot jail again which is not needed
at all as document is already opened. In storages such as WOPI,
such calls can be very expensive. This commit reduces all
such calls made to storage to constant 1 i.e file is downloaded
from storage only during first load by not recreating the storage
object with every load.

Other than that, access token of each session is now correctly
used when interacting with the storage. Earlier, we used to
use the same access token for each upload to storage which means
that irrespective of who clicked the save button, changes to the
document were only made on behalf of one person (of whom the
access token is used). This is fixed now.

Also includes minor cleanup left and right.

Change-Id: Id32702ff02aea4f63b7cc6afa9f62664807bb57d