tdf#103825: Patrol and inform of any to-be-expired sessions
Check all the sessions' access tokens to see if they are about to
expire. If yes, inform loleaflet for the same which should
prompt the user to refresh the session in order to avoid losing
any data. If access_token_ttl value is set to zero or invalid, assume
infinite expiration.

Change-Id: I98c3e47c9b7031e29e002f653d488747b9c17df8