Update patch set 1
Patch Set 1: Code-Review-1

Please are you sure it is correct to return the sendTextFrame()'s return value?  Like, in most cases, sendTextFrame() succeeds, so even for failures, we will return true.

Having said that, the sendFontRendering() return value is completely ignored, we call it as:

  sendFontRendering(buffer, length, tokens);

instead of:

  return sendFontRendering(buffer, length, tokens);

like the others do in ChildSession::_handleInput().  Is that expected?

From what I see, it seems to me that the behaviour after this patch is the same as before that (?)

Patch-set: 1