Update patch set 2
Patch Set 2:

> Looks sensible to me; concerned wrt. performance - but lets see -

The design is very performance conscious. But we still need to unify the other queue with this one and centralize deduplication code. Notice the Message now stores tokens and avoids re-tokenizing on each deduplication pass :)

 > lets get it in and test it in the team call. I'm also nervous about
 > the hand-crafted: std::vector<std::string> tokenize(const char*
 > data, const size_t size) - would love some unit test for that; what

There are unittests in this patch, with very good coverage (corner cases etc.)

 > do we do with leading spaces eg. ? duplicate space separators ?
 > string and/or key parameters that contain a ' ' etc. Really I'd

See unittests in patch for answer :)

 > love us to use JSON for all our messages.

Yes, if for nothing for consistency. Though parsing is significantly more expensive. But with the Message class could be cached.

 > Also - would prolly be worth storing a parsed form of the JSON in
 > the message queue rather than re-parsing as we de-duplicate, but

Yes, doable now that we have a proper Message class, as we do with the tokens, we could do with JSON objects.

 > ... lets see how it goes =) Do we know how long the queue grows to
 > in normal use ?

On a fast connection (i.e. loopback) the queue is not long enough to deduplicate anything. Only with a jitter of 100ms on the socket do deduplication start kicking in and in fact to quite a bit of deduplication of tiles and the cursor movement (and there is no more high latency with too many updates). But! remains to be tested with large group and real connections.

Patch-set: 2
Reviewer: Gerrit User 1000871 <1000871@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Verified=0