Update patch set 1
Patch Set 1:

> Looks ok-ish - but without serious testing under load with lots of
 > users - not hyper-happy about this really. Perhaps with other
 > review of someone really reading the socket code, corner cases etc.
 > etc.

This is a fix to the case described in the commit log.

Notice that there are two distinct, but related, changes here: first the check for error when polling for read fails, but second the check for socket closed from the flags has to be done _before_ the checking for -1, which can also mean timeout. Both of these help us break the loop instead of tight looping and pinning cpu at 100%.

This scenario is not very common, so not urgent to merge in 2.0. Let's wait for more reviews and testing in master (it has been there since Dec 12).

Patch-set: 1
Reviewer: Gerrit User 1000871 <1000871@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Verified=0