Update patch set 1
Patch Set 1:

Noticed this patch now when merging stuff; I wonder, what happens / is supposed to happen if the SELECT_ERROR is only transient - could that be such?  Or is it that once we get SELECT_ERROR, the socket is erroneous for the rest of its lifetime?

Sorry I had a look this late :-( - given that the release is tomorrow, I'd prefer not to merge it today...

Patch-set: 1
Reviewer: Gerrit User 1000068 <1000068@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Code-Review=0