Update patch set 1
Patch Set 1: Code-Review+2 Verified+1

Good idea. Also, having a wall-clock timestamp will make it possible to sort the log file chronologically (as it might not be if the output to it is not flushed after each line by the separate processes that write to it).

Patch-set: 1
Reviewer: Gerrit User 1000062 <1000062@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Verified=+1
Label: Code-Review=+2