Update patch set 1
Patch Set 1: Code-Review-1

Patch requires other fixes already queued (as tagged in their comments) - so we should consider all of these all together; it also changes some strings.
Before this change rendering characters on the server was not tested too heavily: we only rendered one character (and had two views on how it would be rendered in the document: browser + server-side) - after it all of the characters in the grid are server rendered - with some known issues: some of the characters just don't render due to the locale-setting of loolwsd - which is a rather complex open problem.

Personally am not against merging the whole set with proper feature review - and would appreciate a fix for this issue checkout:
https://gerrit.libreoffice.org/gitweb?p=core.git;a=commit;h=af18b7d60f4320fc01b84593af404dbfb4237fcf and
https://gerrit.libreoffice.org/gitweb?p=core.git;a=commit;h=96918103a07ae984befceb4dc579b004216bbc9f (revert)
for the 'obvious' attempt to fix it - which failed.

Patch-set: 1
Reviewer: Gerrit User 1000037 <1000037@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Code-Review=-1