Update patch set 5
Patch Set 5:

Cool, looks good to me now.

However, and you won't like this part much - Michael is working on migrating the admin console to use the new nonblocking sockets, so there would be changes left and right in the admin console, and you would need to rebase your code on the top of it. Lets hold on to this patch till then.

But really good work overall.

Patch-set: 5