Update patch set 1
Patch Set 1: Code-Review-1

Socket.hxx etc. changes look fine; but I would not use 'gzipAllowed' but 'gzip' or somesuch personally in the method signature.

Beyond that - we really don't want to be writing .gz files to the file-system. We want to be storing them in an in-memory stream - and then serving the same data the next time we are asked for it.

[ Ideally we do that read and compression once at startup time too - but that can be a later patch ].

So - promising; you'll need to chase your stream corruption: perhaps some incorrect version of gzip ? or somesuch (?) - and do this in-memory.

Thanks !

Patch-set: 1
Reviewer: Gerrit User 1000037 <1000037@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Code-Review=-1