Update patch set 3
Patch Set 3:

Firstly, from a higher level point of view, I think it would be better if instead of taking the snapshots, you keep the expired Document class objects in the memory. Taking a snapshot of the string representation of the object, IMHO, is bit restrictive I think and it is even repetitive for a single document, eg: you are storing the same pid of the same document everytime in a separate string (in a snapshot).
    When I say restrictive, I mean as the code evolves, string would get overly complicated trying to represent some data in these classes, the relation between them etc.


Yeah, I worked with string representations of the objects because I thought by Admin Console Dashboard perspective (I mean with focus on its viewmodel.. I stored what I would see) but I agree: this way is restrictive so I will change it.

Patch-set: 3
Label: Verified=0