Update patch set 2
Patch Set 2:

(1 comment)

> Hi Pranav,
 > I am working to fix the issues. Sometimes I did overuse of smart
 > pointers because I was looking for avoid memory leaks.

Yes, but if the same purpose can be accomplished by simply declaring a local variable, better use that.

 > I just want to ask you only about Poco::JSON::Object (point 3).
 > With object we may query on a snapshot to get an its specific
 > property eg. myDoc.getHistory().at(pointOfTime).get("memoryDirty").
 > Conversely, with snapshot as a string representation of a json, we
 > always need to split it in some way in order to get some values. I
 > thought object may be more confortable than string, but I will
 > change if we don't need to query them about member variables. Let
 > me know your thought. Thank you!

I did not see anywhere in the code where we are querying that Poco JSON object. We were just converting the JSON string to Poco JSON object and then stringifying it back before sending it to the client. So use of Poco JSON Object looked superfluous to me here with additional performance overhead without providing any significant benefit.

If you think we would be querying the snapshot in future, indeed keeping them in string form is not a good idea. But even in that case, instead of using Poco::JSON::Object, I think it would be better to keep the snapshots in a separate struct or class and then serialize it to a string only when sending it to the client. What do you think ? This approach would be better performance-wise as we do not have to deal with any performance issues related to Poco::JSON::Object parsing of a JSON string and you can obviously query the struct/class anytime you want.

Patch-set: 2