Update patch set 3
Patch Set 3: Code-Review-1

(6 comments)

Regarding your doubt on the bug report about it not working when the admin console is already closed initially, as said on the bug report, you need to store the user name in the View class so that when you query the admin console we are able to get this username.

Regarding the representation of userlist, I thought we would be using an expander or something to show the user list in each row as mentioned in the bug report description, though I am Ok with the current approach for now and you can work on it in subsequent patches, whatever sounds good to you.

Patch-set: 3
Label: Code-Review=-1