Update patch set 1
Patch Set 1:

(3 comments)

Looks good - happy to commit it if it is disabled and the data is heap allocated not on the stack.

But really - I think we need to get into the reading of all the file contents at startup time, and serving them from a hash. Perhaps good to add a new method / constructor on the  "FileServerRequestHandler" that reads all of these files into a static instance and creating that once from LOOLWSD at startup.

Can you do that before we re-enable this (for performance reasons).

Thanks !

Patch-set: 1
Label: Verified=0