Update patch set 1
Patch Set 1:

I can surely do that. but can you please specify how exactly would you like that to be implemented.

I can think of a few ways though:

1. storing(hard coding) the file names that the client will be requesting for, and on start up read each file compress it and save it in-memory in a map(maybe). Then as per the headers we either serve the compressed file directly accessed, or reread the file and serve it (better speed but more memory usage).
2. storing the file names that the client will be requesting for, reading each file, compressing it and writing it with a .gz extension in the file system. If the browser supports compression we read the compressed file and serve it, else we read the normal file and serve that (slower than first option).

And just as you mentioned add this as a new method in "FileServerRequestHandler" and call it from LOOLWSD.

Patch-set: 1
Label: Verified=0