Update patch set 1
Patch Set 1: Code-Review+2 Verified+1

Removes a FIXME, good to have it :-) - thank you!

In general, I am a bit concerned about the complexity of the decisions when to save and when not, hope there is some scope for making it a bit simpler at some stage; but not now of course.

Patch-set: 1
Reviewer: Gerrit User 1000068 <1000068@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Verified=+1
Label: Code-Review=+2