Update patch set 1
Patch Set 1:

> Yes, splitting would be preferred I think.  In particular, I am
 > unhappy about the autolinker - at some stage we'll need to support
 > richtext in the comments too, and I have no idea what consequences
 > it would have with autolinker...

I have ensured that the data which is sent back to LO core upon saving the comment doesn't contain any modifications that autolinker makes to the text. Its only the comment "representation" that is modified (to include <a> tags etc.) just before showing it to the user, the data associated with the comment remains in its pristine form. So, whenever we'll add support for richtext, it should be easy to remove autolinker from the code without making any major changes.

Patch-set: 1